+++ deployments/automation/steps/20211103-srnt-change-source-github-to-bucket.sh
@@ -11,11 +11,19 @@ if [ ${STEP} == "PRE_INIT" ]; then
             if [[ $(yq r "${PWD}/main.tf.json" 'module.pim.product_reference_code') =~ serenity_dev ]]; then
                 # Dev mode
                 yq w -j -P -i ${PWD}/main.tf.json 'module.pim.source' "gcs::https://www.googleapis.com/storage/v1/akecld-terraform-modules/serenity-edition-dev/${PIM_VERSION}//deployments/terraform"
-                yq w -j -P -i ${PWD}/main.tf.json 'module.pim-monitoring.source' "gcs::https://www.googleapis.com/storage/v1/akecld-terraform-modules/serenity-edition-dev/${PIM_VERSION}//deployments/terraform/monitoring"
+                if [[ -z $(yq r "${PWD}/main.tf.json" 'module.pim-monitoring.source') ]]; then
+                  echo "[INFO] No monitoring activated!"
+                else
+                  yq w -j -P -i ${PWD}/main.tf.json 'module.pim-monitoring.source' "gcs::https://www.googleapis.com/storage/v1/akecld-terraform-modules/serenity-edition-dev/${PIM_VERSION}//deployments/terraform/monitoring"
+                fi
             else
                 # Prod mode
                 yq w -j -P -i ${PWD}/main.tf.json 'module.pim.source' "gcs::https://www.googleapis.com/storage/v1/akecld-terraform-modules/serenity-edition/${PIM_VERSION}//deployments/terraform"
-                yq w -j -P -i ${PWD}/main.tf.json 'module.pim-monitoring.source' "gcs::https://www.googleapis.com/storage/v1/akecld-terraform-modules/serenity-edition/${PIM_VERSION}//deployments/terraform/monitoring"
+                if [[ -z $(yq r "${PWD}/main.tf.json" 'module.pim-monitoring.source') ]]; then
+                  echo "[INFO] No monitoring activated!"
+                else
+                  yq w -j -P -i ${PWD}/main.tf.json 'module.pim-monitoring.source' "gcs::https://www.googleapis.com/storage/v1/akecld-terraform-modules/serenity-edition/${PIM_VERSION}//deployments/terraform/monitoring"
+                fi
             fi
         fi
     else