+++ src/Akeneo/ReferenceEntity/back/Infrastructure/Connector/Api/Attribute/CreateOrUpdateAttributeAction.php
@@ -18,6 +18,8 @@ use Akeneo\ReferenceEntity\Infrastructure\Connector\Api\Attribute\JsonSchema\Cre
 use Akeneo\ReferenceEntity\Infrastructure\Connector\Api\Attribute\JsonSchema\Edit\AttributeEditionValidator;
 use Akeneo\ReferenceEntity\Infrastructure\Connector\Api\JsonSchemaErrorsFormatter;
 use Akeneo\Tool\Component\Api\Exception\ViolationHttpException;
+use Oro\Bundle\SecurityBundle\SecurityFacade;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -26,6 +28,8 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\Exception\UnprocessableEntityHttpException;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Routing\Router;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class CreateOrUpdateAttributeAction
@@ -69,6 +73,12 @@ class CreateOrUpdateAttributeAction
     /** @var ValidateAttributePropertiesImmutability */
     private $validateAttributePropertiesImmutability;
 
+    private SecurityFacade $securityFacade;
+
+    private TokenStorageInterface $tokenStorage;
+
+    private LoggerInterface $apiAclLogger;
+
     public function __construct(
         CreateAttributeCommandFactoryRegistry $createAttributeCommandFactoryRegistry,
         FindAttributeNextOrderInterface $attributeNextOrder,
@@ -82,7 +92,10 @@ class CreateOrUpdateAttributeAction
         ValidatorInterface $validator,
         AttributeCreationValidator $jsonSchemaCreateValidator,
         AttributeEditionValidator $jsonSchemaEditValidator,
-        ValidateAttributePropertiesImmutability $validateAttributePropertiesImmutability
+        ValidateAttributePropertiesImmutability $validateAttributePropertiesImmutability,
+        SecurityFacade $securityFacade,
+        TokenStorageInterface $tokenStorage,
+        LoggerInterface $apiAclLogger
     ) {
         $this->createAttributeCommandFactoryRegistry = $createAttributeCommandFactoryRegistry;
         $this->attributeNextOrder = $attributeNextOrder;
@@ -97,10 +110,15 @@ class CreateOrUpdateAttributeAction
         $this->jsonSchemaCreateValidator = $jsonSchemaCreateValidator;
         $this->jsonSchemaEditValidator = $jsonSchemaEditValidator;
         $this->validateAttributePropertiesImmutability = $validateAttributePropertiesImmutability;
+        $this->securityFacade = $securityFacade;
+        $this->tokenStorage = $tokenStorage;
+        $this->apiAclLogger = $apiAclLogger;
     }
 
     public function __invoke(Request $request, string $referenceEntityIdentifier, string $attributeCode): Response
     {
+        $this->denyAccessUnlessAclIsGranted();
+
         try {
             $referenceEntityIdentifier = ReferenceEntityIdentifier::fromString($referenceEntityIdentifier);
             $attributeCode = AttributeCode::fromString($attributeCode);
@@ -272,5 +290,32 @@ class CreateOrUpdateAttributeAction
         ];
 
         return Response::create('', Response::HTTP_NO_CONTENT, $headers);
+    }
+
+    private function denyAccessUnlessAclIsGranted(): void
+    {
+        $acl = 'pim_api_reference_entity_edit';
+
+        if (!$this->securityFacade->isGranted($acl)) {
+            $token = $this->tokenStorage->getToken();
+            if (null === $token) {
+                throw new \LogicException('An user must be authenticated if ACLs are required');
+            }
+
+            $user = $token->getUser();
+            if (!$user instanceof UserInterface) {
+                throw new \LogicException(sprintf(
+                    'An instance of "%s" is expected if ACLs are required',
+                    UserInterface::class
+                ));
+            }
+
+            $this->apiAclLogger->warning(sprintf(
+                'User "%s" with roles %s is not granted "%s"',
+                $user->getUsername(),
+                implode(',', $user->getRoles()),
+                $acl
+            ));
+        }
     }
 }