+++ src/Akeneo/ReferenceEntity/back/Infrastructure/Connector/Api/Record/GetConnectorRecordAction.php
@@ -18,9 +18,13 @@ use Akeneo\ReferenceEntity\Domain\Model\ReferenceEntity\ReferenceEntityIdentifie
 use Akeneo\ReferenceEntity\Domain\Query\Record\Connector\FindConnectorRecordByReferenceEntityAndCodeInterface;
 use Akeneo\ReferenceEntity\Domain\Query\ReferenceEntity\ReferenceEntityExistsInterface;
 use Akeneo\ReferenceEntity\Infrastructure\Connector\Api\Record\Hal\AddHalDownloadLinkToRecordImages;
+use Oro\Bundle\SecurityBundle\SecurityFacade;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\Exception\UnprocessableEntityHttpException;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use Symfony\Component\Security\Core\User\UserInterface;
 
 /**
  * @author Elodie Raposo <elodie.raposo@akeneo.com>
@@ -37,14 +41,26 @@ class GetConnectorRecordAction
     /** @var AddHalDownloadLinkToRecordImages */
     private $addHalLinksToImageValues;
 
+    private SecurityFacade $securityFacade;
+
+    private TokenStorageInterface $tokenStorage;
+
+    private LoggerInterface $apiAclLogger;
+
     public function __construct(
         FindConnectorRecordByReferenceEntityAndCodeInterface $findConnectorRecord,
         ReferenceEntityExistsInterface $referenceEntityExists,
-        AddHalDownloadLinkToRecordImages $addHalLinksToImageValues
+        AddHalDownloadLinkToRecordImages $addHalLinksToImageValues,
+        SecurityFacade $securityFacade,
+        TokenStorageInterface $tokenStorage,
+        LoggerInterface $apiAclLogger
     ) {
         $this->referenceEntityExists = $referenceEntityExists;
         $this->findConnectorRecord = $findConnectorRecord;
         $this->addHalLinksToImageValues = $addHalLinksToImageValues;
+        $this->securityFacade = $securityFacade;
+        $this->tokenStorage = $tokenStorage;
+        $this->apiAclLogger = $apiAclLogger;
     }
 
     /**
@@ -53,6 +69,8 @@ class GetConnectorRecordAction
      */
     public function __invoke(string $referenceEntityIdentifier, string $code): JsonResponse
     {
+        $this->denyAccessUnlessAclIsGranted();
+
         try {
             $recordCode = RecordCode::fromString($code);
             $referenceEntityIdentifier = ReferenceEntityIdentifier::fromString($referenceEntityIdentifier);
@@ -74,5 +92,32 @@ class GetConnectorRecordAction
         $normalizedRecord = current(($this->addHalLinksToImageValues)($referenceEntityIdentifier, [$normalizedRecord]));
 
         return new JsonResponse($normalizedRecord);
+    }
+
+    private function denyAccessUnlessAclIsGranted(): void
+    {
+        $acl = 'pim_api_reference_entity_record_list';
+
+        if (!$this->securityFacade->isGranted($acl)) {
+            $token = $this->tokenStorage->getToken();
+            if (null === $token) {
+                throw new \LogicException('An user must be authenticated if ACLs are required');
+            }
+
+            $user = $token->getUser();
+            if (!$user instanceof UserInterface) {
+                throw new \LogicException(sprintf(
+                    'An instance of "%s" is expected if ACLs are required',
+                    UserInterface::class
+                ));
+            }
+
+            $this->apiAclLogger->warning(sprintf(
+                'User "%s" with roles %s is not granted "%s"',
+                $user->getUsername(),
+                implode(',', $user->getRoles()),
+                $acl
+            ));
+        }
     }
 }