+++ src/Akeneo/ReferenceEntity/back/Infrastructure/Connector/Api/ReferenceEntity/CreateOrUpdateReferenceEntityAction.php
@@ -24,6 +24,8 @@ use Akeneo\ReferenceEntity\Domain\Repository\ReferenceEntityRepositoryInterface;
 use Akeneo\ReferenceEntity\Infrastructure\Connector\Api\JsonSchemaErrorsFormatter;
 use Akeneo\ReferenceEntity\Infrastructure\Connector\Api\ReferenceEntity\JsonSchema\ReferenceEntityValidator;
 use Akeneo\Tool\Component\Api\Exception\ViolationHttpException;
+use Oro\Bundle\SecurityBundle\SecurityFacade;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -31,6 +33,8 @@ use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\UnprocessableEntityHttpException;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Routing\Router;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class CreateOrUpdateReferenceEntityAction
@@ -59,6 +63,12 @@ class CreateOrUpdateReferenceEntityAction
     /** @var ReferenceEntityRepositoryInterface */
     private $referenceEntityRepository;
 
+    private SecurityFacade $securityFacade;
+
+    private TokenStorageInterface $tokenStorage;
+
+    private LoggerInterface $apiAclLogger;
+
     public function __construct(
         ReferenceEntityExistsInterface $referenceEntityExists,
         ValidatorInterface $validator,
@@ -67,7 +77,10 @@ class CreateOrUpdateReferenceEntityAction
         Router $router,
         ReferenceEntityValidator $jsonSchemaValidator,
         FindFileDataByFileKeyInterface $findFileData,
-        ReferenceEntityRepositoryInterface $referenceEntityRepository
+        ReferenceEntityRepositoryInterface $referenceEntityRepository,
+        SecurityFacade $securityFacade,
+        TokenStorageInterface $tokenStorage,
+        LoggerInterface $apiAclLogger
     ) {
         $this->referenceEntityExists = $referenceEntityExists;
         $this->validator = $validator;
@@ -77,10 +90,14 @@ class CreateOrUpdateReferenceEntityAction
         $this->jsonSchemaValidator = $jsonSchemaValidator;
         $this->findFileData = $findFileData;
         $this->referenceEntityRepository = $referenceEntityRepository;
+        $this->securityFacade = $securityFacade;
+        $this->tokenStorage = $tokenStorage;
+        $this->apiAclLogger = $apiAclLogger;
     }
 
     public function __invoke(Request $request, string $referenceEntityIdentifier): Response
     {
+        $this->denyAccessUnlessAclIsGranted();
         try {
             $referenceEntityIdentifier = ReferenceEntityIdentifier::fromString($referenceEntityIdentifier);
         } catch (\Exception $exception) {
@@ -188,5 +205,32 @@ class CreateOrUpdateReferenceEntityAction
             'filePath' => $imageData['filePath'],
             'originalFilename' => $imageData['originalFilename'],
         ];
+    }
+
+    private function denyAccessUnlessAclIsGranted(): void
+    {
+        $acl = 'pim_api_reference_entity_edit';
+
+        if (!$this->securityFacade->isGranted($acl)) {
+            $token = $this->tokenStorage->getToken();
+            if (null === $token) {
+                throw new \LogicException('An user must be authenticated if ACLs are required');
+            }
+
+            $user = $token->getUser();
+            if (!$user instanceof UserInterface) {
+                throw new \LogicException(sprintf(
+                    'An instance of "%s" is expected if ACLs are required',
+                    UserInterface::class
+                ));
+            }
+
+            $this->apiAclLogger->warning(sprintf(
+                'User "%s" with roles %s is not granted "%s"',
+                $user->getUsername(),
+                implode(',', $user->getRoles()),
+                $acl
+            ));
+        }
     }
 }