+++ src/Akeneo/ReferenceEntity/tests/back/Integration/Connector/Api/Context/Distribute/GetConnectorReferenceEntitiesContext.php
@@ -23,7 +23,10 @@ use Akeneo\ReferenceEntity\Domain\Model\ReferenceEntity\ReferenceEntityIdentifie
 use Akeneo\ReferenceEntity\Domain\Query\ReferenceEntity\Connector\ConnectorReferenceEntity;
 use Akeneo\ReferenceEntity\Domain\Repository\ReferenceEntityRepositoryInterface;
 use Akeneo\Tool\Component\FileStorage\Model\FileInfo;
+use AkeneoEnterprise\Test\Acceptance\Permission\InMemory\SecurityFacadeStub;
 use Behat\Behat\Context\Context;
+use Psr\Log\Test\TestLogger;
+use Symfony\Component\HttpFoundation\Response;
 use Webmozart\Assert\Assert;
 
 class GetConnectorReferenceEntitiesContext implements Context
@@ -45,16 +48,37 @@ class GetConnectorReferenceEntitiesContext implements Context
     /** @var array */
     private $referenceEntityPages;
 
+    private ?Response $response = null;
+
+    private SecurityFacadeStub $securityFacade;
+
+    private TestLogger $apiAclLogger;
+
     public function __construct(
         OauthAuthenticatedClientFactory $clientFactory,
         WebClientHelper $webClientHelper,
         InMemoryFindConnectorReferenceEntityItems $findConnectorReferenceEntity,
-        ReferenceEntityRepositoryInterface $referenceEntityRepository
+        ReferenceEntityRepositoryInterface $referenceEntityRepository,
+        SecurityFacadeStub $securityFacade,
+        TestLogger $apiAclLogger
     ) {
         $this->clientFactory = $clientFactory;
         $this->webClientHelper = $webClientHelper;
         $this->findConnectorReferenceEntity = $findConnectorReferenceEntity;
         $this->referenceEntityRepository = $referenceEntityRepository;
+        $this->securityFacade = $securityFacade;
+        $this->apiAclLogger = $apiAclLogger;
+    }
+
+    /**
+     * @BeforeScenario
+     */
+    public function before()
+    {
+        $this->securityFacade->setIsGranted('pim_api_reference_entity_edit', true);
+        $this->securityFacade->setIsGranted('pim_api_reference_entity_list', true);
+        $this->securityFacade->setIsGranted('pim_api_reference_entity_record_edit', true);
+        $this->securityFacade->setIsGranted('pim_api_reference_entity_record_list', true);
     }
 
     /**
@@ -109,6 +133,38 @@ class GetConnectorReferenceEntitiesContext implements Context
                 )
             );
         }
+    }
+
+    /**
+     * @When /^the connector requests all reference entities of the PIM without permission$/
+     */
+    public function theConnectorRequestsAllReferenceEntitiesOfThePIMWithoutPermission()
+    {
+        $this->securityFacade->setIsGranted('pim_api_reference_entity_list', false);
+        $client = $this->clientFactory->logIn('julia');
+
+        $this->response = $this->webClientHelper->requestFromFile(
+            $client,
+            self::REQUEST_CONTRACT_DIR . 'forbidden_list_reference_entities.json'
+        );
+    }
+
+    /**
+     * @Then the PIM notifies the connector about missing permissions for distributing the 7 reference entities of the PIM
+     */
+    public function thePimNotifiesTheConnectorAboutMissingPermissionsForDistributingReferenceEntities()
+    {
+        /**
+         * TODO CXP-923: Assert 403 instead of success & remove logger assertion
+         */
+        $this->webClientHelper->assertJsonFromFile(
+            $this->response,
+            self::REQUEST_CONTRACT_DIR . 'forbidden_list_reference_entities.json'
+        );
+        Assert::true(
+            $this->apiAclLogger->hasWarning('User "julia" with roles ROLE_USER is not granted "pim_api_reference_entity_list"'),
+            'Expected warning not found in the logs.'
+        );
     }
 
     /**